+++ src/common/global-locals.spec.ts
@@ -8,8 +8,10 @@ import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { User } from '../users/user.entity';
 import { getPermissionsFromUser } from '../users/helpers/get-permissions-from-user.helper';
 import { UserPermission } from '../users/user-permission';
+import { getActingUser } from '../users/helpers/get-acting-user.helper';
 
 jest.mock('../users/helpers/get-permissions-from-user.helper');
+jest.mock('../users/helpers/get-acting-user.helper');
 
 let request: DeepMocked<RequestWithAppSession>;
 let response: DeepMocked<Response>;
@@ -78,9 +80,7 @@ function createUserInEnvironment(): User {
     externalIdentifier: '212121',
   });
 
-  request.appSession = {
-    user: user,
-  };
+  (getActingUser as jest.Mock).mockReturnValue(user);
 
   return user;
 }