+++ src/users/helpers/check-user-is-service-owner.helper.ts
@@ -1,0 +1,12 @@
+import { UnauthorizedException } from '@nestjs/common';
+import { User } from '../user.entity';
+
+export function checkUserIsServiceOwner(user: User): void {
+  if (!user) {
+    throw new UnauthorizedException();
+  }
+
+  if (!user.serviceOwner) {
+    throw new UnauthorizedException();
+  }
+}