+++ src/users/users.controller.ts
@@ -29,6 +29,7 @@ import { UserMailer } from './user.mailer';
 import { BackLink } from '../common/decorators/back-link.decorator';
 import { Response } from 'express';
 import { RequestWithAppSession } from '../common/interfaces/request-with-app-session.interface';
+import { getActingUser } from './helpers/get-acting-user.helper';
 
 class UserAlreadyExistsError extends Error {}
 
@@ -51,7 +52,7 @@ export class UsersController {
   @Render('admin/users/index')
   @BackLink('/admin')
   async index(@Req() req: RequestWithAppSession): Promise<IndexTemplate> {
-    const actingUser = req.appSession.user;
+    const actingUser = getActingUser(req);
 
     const users = await (actingUser.serviceOwner
       ? this.usersService.allConfirmed()
@@ -87,7 +88,7 @@ export class UsersController {
   @Post('/admin/users')
   @Permissions(UserPermission.CreateUser)
   async create(@Req() request: RequestWithAppSession, @Res() res: Response) {
-    const actingUser = request['appSession'].user as User;
+    const actingUser = getActingUser(request);
 
     const organisation = actingUser.organisation;